LayoutTests/fast/repaint/implicitly-positioned-block-repaint-simple-line-layout-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('8 541 180 18') is -1
PASS successfullyParsed is true

TEST COMPLETE
this is scrolling and should be repainted.
this should not be repainting
